+++ resources/js/common.js
@@ -1,7 +1,6 @@
 //import '@babel/polyfill';
 import 'bulma/css/bulma.min.css';
 import '@fortawesome/fontawesome-free/js/all.min.js';
-import 'flag-icon-css/css/flag-icon.min.css';
 import '../sass/bulma-theming.scss';
 import '../css/bulma-override.css';
 import '../css/common.css';