+++ webpack.mix.js
@@ -5,7 +5,8 @@ mix.js('resources/js/common.js', 'public/js')
     .js('resources/js/endBorrowing.js', 'public/js')
     .js('resources/js/viewInventory.js', 'public/js')
     .js('resources/js/editInventory.js', 'public/js')
-    .js('resources/js/editUsers.js', 'public/js');
+    .js('resources/js/editUsers.js', 'public/js')
+    .js('resources/js/home.js', 'public/js');
 
 mix.copy('resources/css/index.css', 'public/css')
     .copy('resources/css/new-borrowing.css', 'public/css')
@@ -13,7 +14,8 @@ mix.copy('resources/css/index.css', 'public/css')
     .copy('resources/css/borrowings-history.css', 'public/css')
     .copy('resources/css/view-inventory.css', 'public/css')
     .copy('resources/css/edit-inventory.css', 'public/css')
-    .copy('resources/css/edit-users.css', 'public/css');
+    .copy('resources/css/edit-users.css', 'public/css')
+    .copy('resources/css/flag-icon.min.css', 'public/css');
 
 mix.copyDirectory('resources/favicons', 'public/favicons');