+++ resources/js/common.js
@@ -1,4 +1,3 @@
-//import '@babel/polyfill';
 import 'bulma/css/bulma.min.css';
 import '@fortawesome/fontawesome-free/js/all.min.js';
 import '../sass/bulma-theming.scss';