+++ changelog.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Moved ```public``` folder content to ```resources```
 * Replaced ```Bootstrap``` with ```Bulma```
-* Removing jQuery
+* Removed ```jQuery```
 
 ## [0.14.0] - 2018-10-16